requirements.txt
fastapi
uvicorn
python-socketio
python-dotenv
jinja2
aiofiles
google-generativeai
